Move EvaluationConfiguration to resource
