+++ monitoring/uss_qualifier/resources/netrid/__init__.py
@@ -1,2 +1,3 @@
 from .service_providers import NetRIDServiceProviders
 from .observers import NetRIDObserversResource
+from .evaluation import EvaluationConfigurationResource
