+++ monitoring/uss_qualifier/rid/utils.py
@@ -11,26 +11,9 @@ from monitoring.uss_qualifier.resources.netrid.service_providers import (
 from implicitdict import ImplicitDict, StringBasedTimeDelta
 
 
-class EvaluationConfiguration(ImplicitDict):
-    min_polling_interval: StringBasedTimeDelta = StringBasedTimeDelta("5s")
-    """Do not repeat system observations with intervals smaller than this."""
-
-    max_propagation_latency: StringBasedTimeDelta = StringBasedTimeDelta("10s")
-    """Allow up to this much time for data to propagate through the system."""
-
-    min_query_diagonal: float = 100
-    """Do not make queries with diagonals smaller than this many meters."""
-
-    repeat_query_rect_period: int = 3
-    """If set to a value above zero, reuse the most recent query rectangle/view every this many queries."""
-
-
 class RIDQualifierTestConfiguration(ImplicitDict):
     flight_start_delay: StringBasedTimeDelta = StringBasedTimeDelta("15s")
     """Amount of time between starting the test and commencement of flights"""
-
-    evaluation: EvaluationConfiguration = EvaluationConfiguration()
-    """Settings to control behavior when evaluating observed system data"""
 
     rid_version: RIDVersion
     """Version of remote ID API/standard to use"""
